Fixing #144 ;  not filling buffer while checking for CDATA caused treatening it as comment
